moved maps into the maps.js and brought that component into homepage.js and fixed markers and posted a demo route to first market. ANND added boldness to the nearby market thing.
