+++ App.js
@@ -29,25 +29,25 @@ function App() {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  container2: {
-    flex: 1,
-    width: Dimensions.get('window').width,
-    backgroundColor: '#FA7E61',
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  map: {
-    flex: 1,
-    width: Dimensions.get('window').width,
-    height: Dimensions.get('window').height,
-  },
-});
+// const styles = StyleSheet.create({
+//   container: {
+//     flex: 1,
+//     backgroundColor: '#fff',
+//     alignItems: 'center',
+//     justifyContent: 'center',
+//   },
+//   container2: {
+//     flex: 1,
+//     width: Dimensions.get('window').width,
+//     backgroundColor: '#FA7E61',
+//     alignItems: 'center',
+//     justifyContent: 'center'
+//   },
+//   map: {
+//     flex: 1,
+//     width: Dimensions.get('window').width,
+//     height: Dimensions.get('window').height,
+//   },
+// });
 
 export default App;
